Configure JUnit tests for gradle
